requirements.txt
aiohttp>=3.8.4
aiosignal>=1.3.1
annotated-types>=0.5.0
async-timeout>=4.0.2
attrs>=23.1.0
blinker>=1.6.2
certifi>=2023.5.7
charset-normalizer>=3.1.0
click>=8.1.3
colorama>=0.4.6
dataclasses-json>=0.5.14
Flask>=2.3.2
Flask-Cors>=3.0.10
frozenlist>=1.3.3
greenlet>=2.0.2
idna>=3.4
importlib-metadata>=6.6.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
langchain>=0.0.273
langsmith>=0.0.26
MarkupSafe>=2.1.2
marshmallow>=3.20.1
multidict>=6.0.4
mypy-extensions>=1.0.0
numexpr>=2.8.5
numpy>=1.25.2
openai>=0.27.6
packaging>=23.1
pydantic>=2.3.0
pydantic_core>=2.6.3
python-dotenv>=1.0.0
PyYAML>=6.0.1
regex>=2023.8.8
requests>=2.30.0
six>=1.16.0
SQLAlchemy>=2.0.20
tenacity>=8.2.3
tiktoken>=0.5.1
tqdm>=4.65.0
typing-inspect>=0.9.0
typing_extensions>=4.7.1
urllib3>=2.0.2
Werkzeug>=2.3.4
yarl>=1.9.2
zipp>=3.15.0
